requirements.txt
Django==4.2.14
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
drf-nested-routers==0.93.5  # for nested API endpoints
Pillow==10.4.0
python-dotenv==1.0.1
whitenoise==6.6.0  # for static files in production
gunicorn==21.2.0   # production server
stripe==7.11.0     # payment processing
